# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for estética automotiva dashboard
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  durationMinutes: number;
  maxPerDay: number;
  /** full_day, morning_only, afternoon_only, both_periods */
  period: string;
  description: string;
  color?: string;
}

export interface AvailableDay {
  id: number;
  date: string;
  isAvailable: boolean;
}

export interface AvailableDaysInput {
  month: number;
  year: number;
  dates: string[];
}

export interface Appointment {
  id: number;
  customerName: string;
  customerPhone: string;
  serviceId: number;
  serviceName: string;
  date: string;
  /** morning or afternoon */
  timeSlot: string;
  /** scheduled, completed, cancelled */
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  vehicleModel?: string | null;
  /** @nullable */
  vehiclePlate?: string | null;
}

export interface AppointmentInput {
  customerName: string;
  customerPhone: string;
  serviceId: number;
  date: string;
  notes?: string;
  vehicleModel?: string;
  vehiclePlate?: string;
}

export interface AppointmentUpdate {
  status?: string;
  notes?: string;
  vehicleModel?: string;
  vehiclePlate?: string;
}

export interface AvailableSlots {
  date: string;
  serviceId: number;
  slots: string[];
  canSchedule: boolean;
  /** @nullable */
  reason?: string | null;
}

export interface DashboardSummary {
  todayAppointments: Appointment[];
  weekAppointments: number;
  monthAppointments: number;
  pendingCount: number;
  completedCount: number;
  cancelledCount: number;
  availableDaysThisMonth: number;
  /** @nullable */
  nextAvailableDate?: string | null;
}

export interface BotStatus {
  /** disconnected, connecting, qr_ready, authenticated, ready */
  status: string;
  connected: boolean;
  /** @nullable */
  qrCode?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  lastActivity?: string | null;
}

export interface Conversation {
  phone: string;
  name: string;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
  hasAppointment?: boolean;
}

export interface Message {
  id: string;
  phone: string;
  body: string;
  fromMe: boolean;
  timestamp: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetAvailableDaysParams = {
month: number;
year: number;
};

export type ListAppointmentsParams = {
/**
 * @nullable
 */
date?: string | null;
/**
 * @nullable
 */
month?: number | null;
/**
 * @nullable
 */
year?: number | null;
/**
 * @nullable
 */
status?: string | null;
};

export type GetAvailableSlotsParams = {
date: string;
serviceId: number;
};