# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for estética automotiva dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentUpdate {
  status?: string;
  notes?: string;
  vehicleModel?: string;
  vehiclePlate?: string;
}