# lib/api-zod/src/generated/types/availableSlots.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for estética automotiva dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface AvailableSlots {
  date: string;
  serviceId: number;
  slots: string[];
  canSchedule: boolean;
  /** @nullable */
  reason?: string | null;
}