# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for estética automotiva dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';

export interface DashboardSummary {
  todayAppointments: Appointment[];
  weekAppointments: number;
  monthAppointments: number;
  pendingCount: number;
  completedCount: number;
  cancelledCount: number;
  availableDaysThisMonth: number;
  /** @nullable */
  nextAvailableDate?: string | null;
}