# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for estética automotiva dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  phone: string;
  body: string;
  fromMe: boolean;
  timestamp: string;
}