# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for estética automotiva dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  durationMinutes: number;
  maxPerDay: number;
  /** full_day, morning_only, afternoon_only, both_periods */
  period: string;
  description: string;
  color?: string;
}